+++ src/main/resources/application-default.yaml
@@ -7,7 +7,6 @@ terraware:
   web-app-url: "http://localhost:3000/"
   photo-dir: "/tmp/terraware-server-photos"
   use-test-clock: true
-  allow-admin-ui-for-non-admins: true
   email:
     always-send-to-override-address: true
     subject-prefix: "[DEV]"
